requirements.txt
bcrypt
fastapi[standard]
pyjwt
psycopg2
pydantic
python-dotenv
requests
